feat: autonomous routine (#38)
Added plan for a two ball auton script (movement only, no intake or shooting). Also modified MoveCommand so it has a willFinish to differentiate between auton and teleop commands. Code was tested and it works, but PID might need to be tuned. Also, DO NOT TEST without speed scaling, it goes very quickly.